requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-huggingface==0.1.2
langgraph==0.2.39
gradio==4.44.1
python-dotenv==1.0.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4
pypdf==4.2.0
pydantic==2.9.2
huggingface-hub==0.23.4
transformers>=4.37.0
